requirements.txt
curl-cffi==0.7.4
beautifulsoup4==4.14.2
pycryptodome==3.23.0
m3u8==6.0.0
lxml==6.0.2
pyperclip==1.8.2
